Refactor stores hierarchy

+++ src/store/LibraryStore.ts
@@ -1,19 +1,37 @@
+import { RootStore } from './RootStore'
 import React from 'react'
-import { observable, action } from 'mobx'
+import { observable, action, makeObservable } from 'mobx'
 import libraryDB from '../uitls/clientDB'
 import { BookList, Book } from '../types'
 import * as converter from '../uitls/converter'
 
-export const LibraryStore = () => {
-  const fetchBooksListAction = action(async () => {
-    store.books = await libraryDB.getAllMeta()
-    console.info('book store', store.books)
+type TCurrentBook = null | {
+  info: Book
+  text: string
+  meta: any
+}
+export class LibraryStore {
+  public isAddingBookInProcess: boolean = false
+
+  public books: BookList = []
+  public currentBook: TCurrentBook = null
+  private rootStore: RootStore
+
+  constructor(rootStore: RootStore) {
+    makeObservable(this, {
+      books: observable,
+    })
+    this.rootStore = rootStore
+  }
+  fetchBooksListAction = action(async () => {
+    this.books = await libraryDB.getAllMeta()
+    console.info('book store', this.books)
   })
 
-  const isBookExist = (name: string) =>
-    store.books.find((book: Book) => book.name === name)
+  isBookExist = (name: string | undefined) =>
+    this.books.find((book: Book) => book.name === name)
 
-  const addBookAction = action(async (rawBookText: string, name: string) => {
+  addBookAction = action(async (rawBookText: string, name: string) => {
     const { cover, meta } = converter.getBookPreviewInfo(rawBookText)
     const newBook = {
       name,
@@ -21,75 +39,63 @@ export const LibraryStore = () => {
       cover,
     }
 
-    store.isAddingBookInProcess = true
+    this.isAddingBookInProcess = true
     const book = await libraryDB.addBook(newBook, rawBookText)
-    store.books.push(book)
-    store.isAddingBookInProcess = false
+    this.books.push(book)
+    this.isAddingBookInProcess = false
+
+    this.rootStore.pushMessage('The book has been added')
   })
 
-  const syncBookAction = action(async (meta: Book, body: string) => {
+  syncBookAction = action(async (meta: Book, body: string) => {
     const book = await libraryDB.addBook(meta, body)
-    store.books.push(book)
+    this.books.push(book)
   })
 
-  const removeBookAction = action(async (book: Book) => {
+  removeBookAction = action(async (book: Book | null) => {
+    if (!book) return;
     await libraryDB.delete(book.id)
-    const bookIndex = store.books.indexOf(book)
-    store.books.splice(bookIndex, 1)
+    const bookIndex = this.books.indexOf(book)
+    this.books.splice(bookIndex, 1)
   })
 
-  const updateBookPositionAction = action(
+   updateBookPositionAction = action(
     async (bookId: number, positionElement: string) => {
       await libraryDB.updateBookMeta(bookId, { positionElement })
     }
   )
 
-  const updateBookAction = action(async (bookId: number, bookProps: any) => {
-    const bookInStore = store.books.find((book) => book.id === bookId)
+  updateBookAction = action(async (bookId: number, bookProps: any) => {
+    const bookInStore = this.books.find((book) => book.id === bookId)
     if (bookInStore) {
       Object.assign(bookInStore, bookProps)
     }
     await libraryDB.updateBookMeta(bookId, bookProps)
   })
 
-  const openBookAction = action(async (bookId: number) => {
-    if (store.currentBook!.info!.id === bookId) {
-      store.currentBook.info = await libraryDB.getBookMeta(bookId)
+   openBookAction = action(async (bookId: number) => {
+    if (this.currentBook && this.currentBook.info.id === bookId) {
+      this.currentBook.info =
+        (await libraryDB.getBookMeta(bookId)) || this.currentBook.info
       return Promise.resolve()
+    } else {
+      return Promise.all([
+        libraryDB.getBookMeta(bookId),
+        libraryDB.getBookText(bookId),
+      ]).then((prom) => {
+        const info = prom[0] as Book
+        const rawText = prom[1] || ''
+        const text = converter.parseToInnerBook(rawText)
+        this.currentBook = {
+          info,
+          text,
+          meta: null
+        }
+      })
     }
-    return Promise.all([
-      libraryDB.getBookMeta(bookId),
-      libraryDB.getBookText(bookId),
-    ]).then((prom) => {
-      store.currentBook.info = prom[0]
-      const text = prom[1] || ''
-      store.currentBook.text = converter.parseToInnerBook(text)
-    })
   })
 
-  const fetchBookTextAction = action(
+  fetchBookTextAction = action(
     async (bookId: number) => await libraryDB.getBookText(bookId)
   )
-
-  const store = observable({
-    books: [] as BookList,
-    remoteBooks: [] as BookList,
-    currentBook: {
-      info: {} as Book | undefined,
-      text: '' as string | undefined,
-    },
-    isAddingBookInProcess: false,
-    fetchBooksListAction,
-    fetchBookTextAction,
-    addBookAction,
-    openBookAction,
-    syncBookAction,
-    removeBookAction,
-    updateBookAction,
-    updateBookPositionAction,
-    isBookExist,
-  })
-
-  return store
 }
-export default React.createContext<any>({})
